offline_planner moved from mppi
